+++ BE/iDrop/src/main/java/ifive/idrop/websocket/direction/dto/Direction.java
@@ -1,4 +1,4 @@
-package ifive.idrop.websocket.direction;
+package ifive.idrop.websocket.direction.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
