+++ BE/iDrop/src/main/java/ifive/idrop/websocket/location/LocationWebSocketHandler.java
@@ -10,8 +10,11 @@ import ifive.idrop.jwt.JwtProvider;
 import ifive.idrop.repository.UserRepository;
 import ifive.idrop.util.CustomObjectMapper;
 import ifive.idrop.websocket.PickUpInfoRepository;
-import ifive.idrop.websocket.direction.Direction;
+import ifive.idrop.websocket.location.dto.ChildGeoLocation;
+import ifive.idrop.websocket.location.dto.CurrentPickUp;
+import ifive.idrop.websocket.direction.dto.Direction;
 import ifive.idrop.websocket.direction.NaverDirectionFinder;
+import ifive.idrop.websocket.location.dto.DriverGeoLocation;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
