+++ BE/iDrop/src/main/java/ifive/idrop/websocket/location/dto/ChildGeoLocation.java
@@ -1,4 +1,4 @@
-package ifive.idrop.websocket.location;
+package ifive.idrop.websocket.location.dto;
 
 import lombok.Getter;
